feat: Add comprehensive star rating system
- Add star_rating column to roast_entries table with DECIMAL(2,1) type
- Support 1-5 star scale with 0.5 increments and NULL values for clearing
- Create custom StarRating component with half-star support
- Add StarRatingCard component with clear rating functionality
- Integrate star rating into roast detail page layout
- Display star ratings in roast cards and table views
- Add useStarRating hook for state management and API calls
- Support rating persistence and proper data flow
- Add clear rating functionality with NULL database storage
- Update backend schemas and API endpoints for star rating

+++ backend/schemas.py
@@ -87,7 +87,7 @@ class UpdateRoastRequest(BaseModel):
     weight_after_g: Optional[float] = None
     notes: Optional[str] = None
     tasting_notes: Optional[str] = None
-    star_rating: Optional[int] = None
+    star_rating: Optional[float] = None
 
 
 class UserMachineRequest(BaseModel):
